Vendure provider: split out gql operations, remove unused files
