+++ framework/vendure/lib/fragments/search-result-fragment.ts
