+++ framework/vendure/lib/mutations/log-in-mutation.ts
@@ -1,0 +1,14 @@
+export const loginMutation = /* GraphQL */ `
+  mutation login($username: String!, $password: String!) {
+    login(username: $username, password: $password) {
+      __typename
+      ... on CurrentUser {
+        id
+      }
+      ... on ErrorResult {
+        errorCode
+        message
+      }
+    }
+  }
+`
