+++ framework/vendure/lib/mutations/log-out-mutation.ts
@@ -1,0 +1,7 @@
+export const logoutMutation = /* GraphQL */ `
+  mutation logout {
+    logout {
+      success
+    }
+  }
+`
