+++ framework/vendure/lib/mutations/remove-order-line-mutation.ts
@@ -1,0 +1,15 @@
+import { cartFragment } from '../fragments/cart-fragment'
+
+export const removeOrderLineMutation = /* GraphQL */ `
+  mutation removeOrderLine($orderLineId: ID!) {
+    removeOrderLine(orderLineId: $orderLineId) {
+      __typename
+      ...Cart
+      ... on ErrorResult {
+        errorCode
+        message
+      }
+    }
+  }
+  ${cartFragment}
+`
