+++ framework/vendure/lib/mutations/sign-up-mutation.ts
@@ -1,0 +1,14 @@
+export const signupMutation = /* GraphQL */ `
+  mutation signup($input: RegisterCustomerInput!) {
+    registerCustomerAccount(input: $input) {
+      __typename
+      ... on Success {
+        success
+      }
+      ... on ErrorResult {
+        errorCode
+        message
+      }
+    }
+  }
+`
