+++ framework/vendure/lib/normalize.ts
@@ -1,73 +1,5 @@
 import { Cart, Product } from '@commerce/types'
-import update from '../lib/immutability'
 import { CartFragment, SearchResultFragment } from '../schema'
-
-function normalizeProductOption(productOption: any) {
-  const {
-    node: {
-      entityId,
-      values: { edges },
-      ...rest
-    },
-  } = productOption
-
-  return {
-    id: entityId,
-    values: edges?.map(({ node }: any) => node),
-    ...rest,
-  }
-}
-
-export function normalizeProduct(productNode: any): Product {
-  const {
-    entityId: id,
-    productOptions,
-    prices,
-    path,
-    id: _,
-    options: _0,
-  } = productNode
-
-  return update(productNode, {
-    id: { $set: String(id) },
-    images: {
-      $apply: ({ edges }: any) =>
-        edges?.map(({ node: { urlOriginal, altText, ...rest } }: any) => ({
-          url: urlOriginal,
-          alt: altText,
-          ...rest,
-        })),
-    },
-    variants: {
-      $apply: ({ edges }: any) =>
-        edges?.map(({ node: { entityId, productOptions, ...rest } }: any) => ({
-          id: entityId,
-          options: productOptions?.edges
-            ? productOptions.edges.map(normalizeProductOption)
-            : [],
-          ...rest,
-        })),
-    },
-    options: {
-      $set: productOptions.edges
-        ? productOptions?.edges.map(normalizeProductOption)
-        : [],
-    },
-    brand: {
-      $apply: (brand: any) => (brand?.entityId ? brand?.entityId : null),
-    },
-    slug: {
-      $set: path?.replace(/^\/+|\/+$/g, ''),
-    },
-    price: {
-      $set: {
-        value: prices?.price.value,
-        currencyCode: prices?.price.currencyCode,
-      },
-    },
-    $unset: ['entityId'],
-  })
-}
 
 export function normalizeSearchResult(item: SearchResultFragment): Product {
   return {
