+++ framework/vendure/lib/queries/active-customer-query.ts
@@ -1,0 +1,10 @@
+export const activeCustomerQuery = /* GraphQL */ `
+  query activeCustomer {
+    activeCustomer {
+      id
+      firstName
+      lastName
+      emailAddress
+    }
+  }
+`
