+++ framework/vendure/lib/queries/get-all-product-paths-query.ts
@@ -1,0 +1,9 @@
+export const getAllProductPathsQuery = /* GraphQL */ `
+  query getAllProductPaths($first: Int = 100) {
+    products(options: { take: $first }) {
+      items {
+        slug
+      }
+    }
+  }
+`
