+++ framework/vendure/lib/queries/get-all-products-query.ts
@@ -1,0 +1,12 @@
+import { searchResultFragment } from '../fragments/search-result-fragment'
+
+export const getAllProductsQuery = /* GraphQL */ `
+  query getAllProducts($input: SearchInput!) {
+    search(input: $input) {
+      items {
+        ...SearchResult
+      }
+    }
+  }
+  ${searchResultFragment}
+`
