+++ framework/vendure/lib/queries/get-cart-query.ts
@@ -1,0 +1,10 @@
+import { cartFragment } from '../fragments/cart-fragment'
+
+export const getCartQuery = /* GraphQL */ `
+  query activeOrder {
+    activeOrder {
+      ...Cart
+    }
+  }
+  ${cartFragment}
+`
