+++ framework/vendure/lib/queries/get-collections-query.ts
@@ -1,0 +1,21 @@
+export const getCollectionsQuery = /* GraphQL */ `
+  query getCollections {
+    collections {
+      items {
+        id
+        name
+        description
+        slug
+        productVariants {
+          totalItems
+        }
+        parent {
+          id
+        }
+        children {
+          id
+        }
+      }
+    }
+  }
+`
