+++ framework/vendure/lib/queries/search-query.ts
@@ -1,0 +1,13 @@
+import { searchResultFragment } from '../fragments/search-result-fragment'
+
+export const searchQuery = /* GraphQL */ `
+  query search($input: SearchInput!) {
+    search(input: $input) {
+      items {
+        ...SearchResult
+      }
+      totalItems
+    }
+  }
+  ${searchResultFragment}
+`
