+++ framework/vendure/product/get-product.ts
@@ -1,48 +1,7 @@
 import { Product } from '@commerce/types'
 import { getConfig, VendureConfig } from '../api'
-import { GetProductQuery } from '@framework/schema'
-
-export const getProductQuery = /* GraphQL */ `
-  query getProduct($slug: String!) {
-    product(slug: $slug) {
-      id
-      name
-      slug
-      description
-      assets {
-        id
-        preview
-        name
-      }
-      variants {
-        id
-        priceWithTax
-        currencyCode
-        options {
-          id
-          name
-          code
-          groupId
-          group {
-            id
-            options {
-              name
-            }
-          }
-        }
-      }
-      optionGroups {
-        id
-        code
-        name
-        options {
-          id
-          name
-        }
-      }
-    }
-  }
-`
+import { GetProductQuery } from '../schema'
+import { getProductQuery } from '../lib/queries/get-product-query'
 
 async function getProduct({
   query = getProductQuery,
